Add render method

+++ src/render/Geometry.ts
@@ -8,7 +8,7 @@ export class Geometry {
   // be used inside a game engine?
   attributes: Map<string, AttributeEntry<Float32Array>> = new Map();
 
-  elements: AttributeEntry<Uint32Array> | null = null;
+  elements: AttributeEntry<Uint16Array> | null = null;
 
   getBuffer(name: string): Float32Array | null {
     const entry = this.attributes.get(name);
@@ -26,12 +26,12 @@ export class Geometry {
     }
   }
 
-  getElementsBuffer(): Uint32Array | null {
+  getElementsBuffer(): Uint16Array | null {
     if (this.elements == null) return null;
     return this.elements.array;
   }
 
-  setElementsBuffer(value: Uint32Array): void {
+  setElementsBuffer(value: Uint16Array): void {
     if (this.elements == null) {
       this.elements = { array: value, version: 0 };
     } else {
